libft.h done et all prototype OK

+++ ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: jsavard <jsavard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 13:36:42 by jsavard           #+#    #+#             */
-/*   Updated: 2022/10/18 15:25:40 by jsavard          ###   ########.fr       */
+/*   Updated: 2022/10/19 14:26:02 by jsavard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	int	i;
